% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dotaro_decks.R
\name{dotaro_decks}
\alias{dotaro_decks}
\title{Return piecepackr configurations of Dotaro Decks}
\usage{
dotaro_decks(..., round = TRUE, border = TRUE)
}
\arguments{
\item{...}{Ignored}

\item{round}{Should the card's corners be rounded}

\item{border}{Should the card have a black border}
}
\value{
A list of \code{\link[piecepackr:pp_cfg]{piecepackr::pp_cfg()}} objects.
}
\description{
\code{dotaro_decks()} returns a list of piecepackr configurations
}
\examples{
decks <- dotaro_decks()
if (piecepackr:::device_supports_unicode()) {
  cfg <- decks$dotaro_full_traditional
  grid::grid.newpage()
  piecepackr::grid.piece("card_face", suit = 4, rank = 4, cfg = cfg)
}
if (piecepackr:::device_supports_unicode()) {
  grid::grid.newpage()
  piecepackr::grid.piece("card_face", suit = 8, rank = 14, cfg = cfg)
}
if (piecepackr:::device_supports_unicode()) {
  cfg <- decks$dotaro_full_number
  grid::grid.newpage()
  piecepackr::grid.piece("card_face", suit = 4, rank = 5, cfg = cfg)
}
if (piecepackr:::device_supports_unicode()) {
  grid::grid.newpage()
  piecepackr::grid.piece("card_face", suit = 9, rank = 9, cfg = cfg)
}
if (piecepackr:::device_supports_unicode()) {
  cfg <- decks$dotaro_full_fool
  grid::grid.newpage()
  piecepackr::grid.piece("card_face", suit = 2, rank = 2, cfg = cfg)
}
}
