% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/save_manual.R
\name{save_manual}
\alias{save_manual}
\title{Save manual of Dotaro Deck}
\usage{
save_manual(
  filename = "dotaro_manual.pdf",
  ...,
  variant = c("french_bw", "french_color", "hybrid"),
  size = c("letter", "a4"),
  quietly = TRUE
)
}
\arguments{
\item{filename}{Filename}

\item{...}{Ignored}

\item{variant}{One of \code{"french_bw"}, \code{"french_color"}, or \code{"hybrid"}.
Which suit style and color scheme to render
the manual's example figures in.  See \code{\link[=dotaro_deck_options]{dotaro_deck_options()}}.}

\item{size}{Either \code{"letter"} or \code{"a4"}.}

\item{quietly}{Whether to hide document compilation output.}
}
\value{
The filename invisibly.  As a side effect creates a pdf file.
}
\description{
\code{save_manual()} saves a pdf manual explaining the Dotaro Deck, knit from
the \code{inst/Rtex} Rtex files via \code{knitr} and compiled with \code{xelatex}.
}
