% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/save_print_and_play.R
\name{save_print_and_play}
\alias{save_print_and_play}
\title{Save print-and-play file of Dotaro Deck}
\usage{
save_print_and_play(
  filename = "dotaro.pdf",
  ...,
  size = c("letter", "a4"),
  title = "Dotaro Deck"
)
}
\arguments{
\item{filename}{Filename}

\item{...}{Ignored}

\item{size}{Either \code{"letter"} or \code{"A4"}.}

\item{title}{pdf metadata title.}
}
\value{
The filename invisible.  As a side effect creates a pdf file.
}
\description{
\code{save_print_and_play()} saves a print-and-play file
}
